MODE-1364 Corrected use of reference in join condition
The indexes were not storing the value of REFERENCE or WEAK REFERENCE properties, and thus
null was returned in the tuples for these records. Consequently, the join could not be performed
with null values.

Added several test cases that replicated this problem and now verify the updated code
exhibits the correct behavior.