MODE-1366 Error parsing JCR-SQL2 queries using square brackets for quoting
The JCR-SQL2 grammar uses square brackets for quoting identifiers, node names
and node paths. However, path literals may also have square brackets around
the same-name-siblings. The JCR-SQL2 parser was not able to properly parse
quoted paths with SNS indexes.

The fix is to adjust the tokenizer to count the number of open and close
square brackets, and to stop the quoted string token when the all of the opened
square brackets were closed. Additional unit tests were added to verify the
tokenizer and JCR-SQL2 parser behavior.

All unit and integration tests pass with these changes.