MODE-1336 Fixed reading of the same file in the binary store by multiple concurrent readers
The old code created new input streams over the same file channel, which means that concurrent reading was not possible (each reader would see a different position in the stream). The solution was to not use the same file channel and make sure each reader has its own (dedicated) stream.