MODE-1368 Removed all legacy modules no longer needed in 3.x
ModeShape 3.x will not need a number of the 2.x modules. In particular:
- since 3.x will only have an AS7 kit, the AS5 or AS6 artifacts were removed
- all the connectors were removed, since they're no longer used
- the connector benchmark tests module was removed, replaced by our new
  performance test suite
- the JPA DDL generator utility has been removed
- the 'modeshape-graph', 'modeshape-repository', 'modeshape-search-lucene'
  and 'modeshape-clustering' modules have all been removed, since the new
  'modeshape-jcr' module no longer uses them
- the DocBook modules were removed and are replaced by the Confluence space
- the two JDBC modules were moved out of the 'utils' directory to top-level modules

The build still works, but not all components have been included in the build.
This is because the query functionality doesn't yet work, so quite a few web
and JDBC driver modules all depend on this.

The assembly profile has not yet been changed or corrected.