MODE-1343 Changed descriptions in POM files, removed unused code
Several of the descriptions in the POM files of the sequencer modules were changed
to be more consistent and easily readable in the Maven results. The old descriptions
came from 2.x and have not been changed for quite a while. But the new descriptions
seem to work better and provide more readable Maven output. Also corrected the parent
coordinates in the text connector.

Removed several of the 'extensions/modeshape-sequencer-*' directories that were
recently converted. These older directories are no longer used.