MODE-1400 Added error/warning checks when registering node types via CND files
Previously, the CND files were being read but any problems during the reading were not reported and were simply
ignored. Some errors cause the reader to abort reading and to have successfully read in no node types, so
registration appeared to do nothing while hiding the errors. With these changes, the errors and warnings
are included in a thrown exception; if there are no errors and only warnings, the warnings are simply logged.

All unit and integration tests pass.