MODE-1402 Changed repository configuration format to simplify sequencer configuration
Made a couple of changes to the repository configuration JSON Schema to simplify things
and remove unused or ancillary items:
- The sequencer, text extractor, and authentication provider components no longer have a
"name" field, since it was not much different than "description"
- The sequencer, text extractor, and authentication provider components have a "type"
field that is a superset of the older "type" and "classname" fields. The "type" field
is now required.
- The sequencer components no longer have a "pathExpression" and "pathExpressions" fields;
instead, there's just "pathExpressions" whose values are always arrays. Note that
"pathExpressions" is now a required field.

Lots of configuration files for the sequencer modules were changed (as was the online
documentation). Quite a few other classes needed to be changed to support the removal
of the "name" field from Component and Sequencers.

All unit and integration tests pass.