MODE-1307 Added Content-Type HTTP header to REST examples
The examples in the RESTful Service chapter did not show including the "Content-Type" HTTP
header. The service will correctly return JSON if this header or "Content-Type" header is
supplied in the request. However, an error will result if a different content-Type header
is provided.

This change only affects the Reference Guide.