MODE-1379 MODE-1426 1427 Upgraded Infinispan, JGroups and PicketBox
The dependency on Infinispan was upgraded to 5.1.2.FINAL, which is the same version used by
JBoss AS 7.1.1.Final. The JGroups version was also changed to match that used by Infinispan
5.1.2.FINAL (and JBoss AS 7.1.1.Final). Also, the PicketBox version was upgraded to the
same version used in JBoss AS 7.1.1.Final.

The Infinispan upgrade required some changes to how the DeltaAware-related code works, since
that framework was changed between 5.1.0.BETA3 and 5.1.0.FINAL (though this was unchanged
since then).

All current unit and integration tests pass.