MODE-1405 Corrected how JCR import handles protected content
Corrected the behavior of the JCR import logic to always import protected content.
The whole point of importing a previously-exported file is to recover all of the
content the way it was, including the 'jcr:created', 'jcr:createdBy' and other
protected properties. (Although note that the 'jcr:lastModified' and 'jcr:lastModifiedBy'
properties will still be updated during the import.)

Added a new unit test to verify that the import now works correctly with protected content.