MODE-1439 Added new JndiRepositoryFactory implementation in ModeShape's API
Added a new JNDI-based RepositoryFactory that looks for only "jndi:"-type URLs, and is
aware of both Repository instances and (ModeShape-specific) Repositories instances.
This allowed simplification of the web library used in the RESTful and WebDAV services,
since they only need to specify the normal RepositoryFactory parameters in the
web.xml.

All unit and integration tests pass.