MODE-1435 Refactored ModeShape subsystem for AS7
Redesigned the XML schema for the ModeShape subsystem, incorporating all of the features
that we'll want to have included in the configuration. Also changed significantly how
the XML elements and attributes are processed and written, and separated them from the
AttibuteDefinitions used in the model (which are now used to parse the XML literals into
the model values and marshal the model values into XML). This new design was patterned
after Infinispan's subsystem and relies more upon code provided by DMR and AS7.

There are operations for adding and removing sequencers, index stores and binary stores;
other operations will be added in later commits.

Finally, a number of additional tests were added to more thoroughly test the subsystem
and various configurations, including reading and writing the XML configurations
and comparing to expected JSON configurations.

Much of this change was patterned after the Infinispan subsystem.