MODE-1449 Renamed internal Binary interface to BinaryValue
ModeShape had 3 Binary interfaces, including the javax.jcr.Binary and the org.modeshape.jcr.api.Binary
extension in the public API. The additional org.modeshape.jcr.value.Binary interface is really an internal
value, but having it share the same name as the other two is more confusing and can lead to programming
errors.

This internal interface was renamed to BinaryValue. A significant amount of code needed to be changed,
which shows how much the internal interface is used within our codebase.

All tests pass with these changes.