MODE-1422 - Fixed the invalid ReferenceIntegrityException in relation with SNS.
The original issue was caused by the fact that SNS location changes are performed directly against the cached nodes, in the transient session, while the query matches location from the indexes against the transient locations, causing invalid nodes to be returned by the query results.

The fix was to make the query results locate the node by the UUID first (if such information is available) and only as a fallback, use the location directly.