MODE-1435 Upgraded MongoDB Java driver used to test JSON/BSON I/O
The MongoDB driver is used in some of Schematic's test cases to validate the I/O of the JSON and BSON readers and writers. The version of the MongoDB driver was upgraded to the latest; since this is only used for tests, it has no influence on any non-test artifacts.