MODE-1484 Added more error handling to file system connector
The file system connector was not properly handling the case when "java.io.File.list(FilenameFilter)"
returns null, which it can sometimes do if there's an I/O error. Now, if this happens, the method
call is retried up to 4 more times; after that, if the call still returns null, an error message will
be logged and no node will be created for the problematic file.

All unit and integration tests pass with these changes.