MODE-1492 Changed JCR TCK artifact version to 2.5.0
This version includes several fixes for the TCK unit tests, including:
- https://issues.apache.org/jira/browse/JCR-2662 Journaled Observation tests should check for repository support ...
- https://issues.apache.org/jira/browse/JCR-3300 Tests should check for repository support ...
- https://issues.apache.org/jira/browse/JCR-3307 Activities tests should check for repository support ...

However, two issues are still outstanding in 2.5.0:
- https://issues.apache.org/jira/browse/JCR-2666 (Patch submitted 5/25/12, but test fails when running Jackrabbit Core) Restore tests assert incorrectly
- https://issues.apache.org/jira/browse/JCR-3313 (Patch submitted 5/24/12) QOM test asserts are too restrictive

In addition to changing the version number of the "{{org.apache.jackrabbit.jcr-tests}}" dependency, we also had to add some properties so that the new tests used node types that had required but not auto-created properties. Once this was done, all the same TCK tests passed with no problems.

All our unit tests pass with these changes.