MODE-1498 Added initial support for JTA transactions
With this commit, container-managed transactions and even user-managed transactions (including both local
and distributed) should work, per Chapter 21 of the JSR-283 specification. Even when using XA transactions,
ModeShape works with the transactions (via Synchronizations) and Infinispan automatically registers the
cache as a resource in the transaction.

However, this commit does not yet change JcrSession to implement XAResource. We need to make this change
per Section 21.4 of the specification so that users can directly enlist the ModeShape Session instance
into the transaction. Implementing XAResource will be completed in a subsequent commit.

The build runs successfully with these changes.