MODE-1489 ModeShape's JCR Session now implements XAResource (for 'auto' transactions)
When the 'auto' transaction mode is enabled (this is the default setting), ModeShape's JCR Session implements XAResource, allowing JCR clients to explicitly enlist the Session as a resource in the transaction. In all other ways, the new JcrXaSession is identical to the JcrSession class. The JcrXaSession class delegates all calls to the XAResource of the Infinispan cache, and (as with the other JTA-related changes) ModeShape participates in the transaction via javax.transaction.Synchronization mechanism.

When the 'none' transaction mode is used, ModeShape's session does not implement XAResource.