MODE-1508 Removed 'provided' scopes that cause JavaDoc issues
The Maven JavaDoc plugin requires 'compile' scope to find referenced or linked classes,
and the recently-changed 'provided' scope for the JBoss AS7 subsystem meant the
JavaDoc could not be successfully built for the AS7 subsystem. Thus we needed to
change the 'provided' scopes back to 'compile' for the AS7 subsystem.

(Also, I couldn't find the AS7 JavaDocs to see if linking to them changed the result.)