MODE-1517 Added support for descriptions (comments) in JSON configuration
The JSON Schema used for ModeShape repository configuration files allows has "description" fields
in most of the sections of the JSON document, and these can be used to describe the different parts
of the document (much the same was as comments). ModeShape doesn't use any of these fields
for any purposes.