MODE-1515 - Extracted AS7 "common" module and fixed the dependencies between modules.
Extracting the common module did bring up a small issue in that i18n resources from other modules (e.g. main) could not be loaded. The reason what that the ClasspathLocalizationRepository always used its own CL to load the .properties file, which doesn't work with the AS7 classloading mechanism. The solution was to simplify the loading  and always use the class loader of the i18n class which we're trying to localize.