MODE-1516 Corrected repository when configuration is updated
When a repository configuration was updated (e.g., when the sequencers are added to the repository,
even in the configuration file), the AS7-specific Environment implementation
was being lost, causing a problem finding the correct cache container an cache. The problem
was that JcrRepository was not correctly transferring the Environment object to the
new/updated RepositoryConfiguration. Once this was done, then everything seems to work
just fine.