MODE-1514 Renamed JcrEngine to ModeShapeEngine
We're trying to position ModeShape first and foremost as a distributed, hierarchical, transactional, and consistent data store
that happens to use the JCR API for its public API. Therefore, ModeShape's engine isn't so much a "JCR engine" as it is
just a "ModeShape engine". Thus, the renaming.

All unit and integration tests pass with these changes.