MODE-1522 Changed the sequencer portion of the JSON configuration
Previously the sequencers were defined in the JSON configuration using an array of nested
documents. However, this made it difficult to easily identify and update a particular sequencer
configuration. And even though the names were intended to be unique, using an array did
not enforce that constraint. Finally, the sequencer configurations are not ordered.

Now, the 'sequencing/sequencers' field in the JSON configuration files are documents,
with the field name containing the unique name of the sequencer. The value of these
fields are each the sequencer configuration for the given name.

An error in the JSON Schema validation logic was discovered and corrected. Note that very
few of the changes involved the production code; most changes were in fact for test code
or configurations.