MODE-1515 - Updated classloading mechanism used for components
Since AS7 uses modules dedicated classloaders, it was necessary to update the way in which components (sequencers, extractors etc) are loaded. The solution was to move the classloader creation to the Environment (with the dedicated LocalEnvironment and RepsitoryService implementations)  and to use a new class loader implementation: DelegatingClassLoader. This implementation uses an inner list of class loader and delegates each operation to them. In the case of AS7, there is a delegate class loader which uses the ModuleLoader to find the classloader for each module.

The changes also contain an update to the common package: the component package and most of its unused classes have been removed. The classes which are still used have been moved to .util.