MODE-1526 Correct persistence functionality
The DocumentExternalizer was checking the number of bytes available for reading without blocking, and this seemed to always return 0 bytes. The DocumentExternalizer was then immediately returning without actually reading any of the content; obviously this was wrong. Now, with some minor changes to the BSON reader/writer, the DocumentExternalizer is merely always just reading the BSON document.

This appears to have fixed the persistence issue, and the recently-added unit test now runs (after some modifications). Note that we're still apparently having issues with the index (native) write locks being kept after shutdown; this will be worked under MODE-1512.