MODE-1512 Corrected tests to show that the locks are being removed
Recent changes added a close mechanism to ensure that the Hibernate Search
SearchFactory is properly closed when the repository is closed. This mechanism
was refined a bit, and the tests were refactored to pull the test that
starts, uses, stops, and restarts the repository (by creating a new JcrRepository)
instance into a separate test class to reduce the number of repositories
that are used.

All unit and integration tests pass with these changes.