MODE-1536 Corrected use of System.nanoTime() for all platforms
Different platforms implement System.nanoTime() differently, and it's possible that computing
the difference between two subsequent calls to the method will be a negative number. Therefore,
changed all the places where we're using it to use the absolute value of the difference.
On most platforms, this won't result in any change, but on some Windows platforms this
fix should prevent computing negative durations.

All unit and integration tests pass.