MODE-1538 Added ability to cancel queries
Added a Query.cancel() method that can be used to cancel currently running queries.
Note that because the Query.execute() method blocks, canceling a query must be done
via a separate thread. This does add some complexity to clients, but it provides
a very useful capability to cancel long-running (or perhaps runaway) queries.

All tests pass with these changes.