MODE-1525 Adding a node without specifying primary type incorrectly assumed "nt:unstructured"
If a parent has a child node definition with a default primary node type, and a child
(with the correct name to match the child node definition) is added without specifying
the primary type, the child incorrectly was assigned a primary type of "nt:unstructured".
This was because the "AbstractJcrNode.addNode(String)" method incorrectly hard-coded
the node type to "nt:unstructured" rather than determining the child's node type from
the parent's node type(s). Simply removing this assumption fixes the problem.

Two new tests were added to replicate this problem and verify the changes do correct the
problem.

All unit and integration tests pass with these changes.